torch-qa/demo1/requirements.txt
requests
bs4
markdown
html2text
pymysql
uvicorn
pydantic
#celery
#redis
python-dotenv
flask
flask_cors